Test fails: RpcAPI.getResource does now count query results starting at 1.
svn path=/trunk/eXist-1.0/; revision=1368


Former-commit-id: 620827f3475cc7dd7e877cb47dc242e78255209a

+++ src/org/exist/xmlrpc/test/XmlRpcTest.java
@@ -170,7 +170,7 @@ public class XmlRpcTest extends TestCase {
         
         params.clear();
         params.addElement(handle);
-        params.addElement(new Integer(0));
+        params.addElement(new Integer(1));
         params.addElement(options);
         byte[] item = (byte[]) xmlrpc.execute( "retrieve", params );
         assertNotNull(item);
